requirements.txt
bcrypt==4.3.0
fastapi~=0.121.2
aiogram~=2.25.1
passlib~=1.7.4
uvicorn~=0.38.0
typing_extensions~=4.15.0
aiohttp~=3.13.3
dominate~=2.9.1
PyYAML~=6.0.3
Jinja2~=3.1.6
pydantic~=2.11.10